requirements.txt
pandas>=1.3.0
numpy>=1.21.0
PyYAML>=5.4.1
chardet>=4.0.0
openpyxl>=3.0.7